+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Please add changes to "master", preferably ordered by their significance. (Most 
  - Brand colors used in several new places, for consistent styling.
  - Users can now set SQL box summaries.
  - SQL boxes can optionally persist their outputs.
+ - Train a decision tree classification model can use String attributes.
 
 ### 2.0.0